progress refactoring texture handling
