+++ libs/opengl/src/registerAllClasses.cpp
@@ -54,7 +54,7 @@ MRPT_INITIALIZER(registerAllClasses_mrpt_opengl)
 	registerClass(CLASS_ID(CSetOfObjects));
 	registerClass(CLASS_ID(CSetOfTriangles));
 	registerClass(CLASS_ID(CSetOfTexturedTriangles));
-	registerClass(CLASS_ID(CSkyBox);
+	registerClass(CLASS_ID(CSkyBox));
 	registerClass(CLASS_ID(CSimpleLine));
 	registerClass(CLASS_ID(CSphere));
 	registerClass(CLASS_ID(CText));
